Add API to delete a chat feedback

+++ ApplensBackend/Services/CognitiveSearchService/CognitiveSearchAdminService.cs
@@ -1,5 +1,6 @@
 ﻿using AppLensV3.Models;
 using Azure.Search.Documents.Models;
+using Microsoft.Azure.Cosmos.Serialization.HybridRow;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
@@ -11,6 +12,8 @@ namespace AppLensV3.Services.CognitiveSearchService
     public interface ICognitiveSearchAdminService
     {
         Task<bool> AddDocuments(List<CognitiveSearchDocument> documents, string indexName);
+        Task<Tuple<bool, List<string>>> DeleteDocuments(List<string> documentIds, string indexName, string idColumnName = "Id");
+        Task<Tuple<bool, List<string>>> DeleteDocuments(List<CognitiveSearchDocument> documents, string indexName);
         Task<bool> DeleteIndex(string indexName);
         Task<bool> CreateIndex(string indexName);
         Task<List<string>> ListIndices();
@@ -24,20 +27,46 @@ namespace AppLensV3.Services.CognitiveSearchService
             _baseService = baseService;
         }
 
-        public async Task<bool> DeleteDocuments(List<CognitiveSearchDocument> documents, string indexName)
+        public async Task<Tuple<bool, List<string>>> DeleteDocuments(List<CognitiveSearchDocument> documents, string indexName)
         {
+            if (documents?.Count < 1)
+            {
+                return new Tuple<bool, List<string>>(true, new List<string>());
+            }
+
             var searchClient = await _baseService.GetIndexClientForAdmin(indexName);
             IndexDocumentsResult result = await searchClient.DeleteDocumentsAsync(documents.Select(document => CreateDocumentModel(document)));
-            return !(result.Results.Any(r => r.Succeeded == false) == true);
+            List<string> deletedDocuments = result.Results.Where(r => r.Succeeded == true).Select(r => r.Key).ToList();
+            return new Tuple<bool, List<string>>(!(result.Results.Any(r => r.Succeeded == false) == true), deletedDocuments ?? new List<string>());
+        }
+
+        public async Task<Tuple<bool, List<string>>> DeleteDocuments(List<string> documentIds, string indexName, string idColumnName = "Id")
+        {
+            if (documentIds?.Count < 1)
+            {
+                return new Tuple<bool, List<string>>(true, new List<string>());
+            }
+
+            var searchClient = await _baseService.GetIndexClientForAdmin(indexName);
+            IndexDocumentsResult result = await searchClient.DeleteDocumentsAsync(idColumnName, documentIds);
+            List<string> deletedDocuments = result.Results.Where(r => r.Succeeded == true).Select(r => r.Key).ToList();
+            return new Tuple<bool, List<string>>(!(result.Results.Any(r => r.Succeeded == false) == true), deletedDocuments ?? new List<string>());
         }
 
         public async Task<bool> AddDocuments(List<CognitiveSearchDocument> documents, string indexName)
         {
-            IndexDocumentsBatch<CognitiveSearchDocumentWrapper> batch = IndexDocumentsBatch.Create(
+            if (documents?.Count > 0 && !string.IsNullOrWhiteSpace(indexName))
+            {
+                IndexDocumentsBatch<CognitiveSearchDocumentWrapper> batch = IndexDocumentsBatch.Create(
                 documents.Select(document => IndexDocumentsAction.Upload(CreateDocumentModel(document))).ToArray());
-            var searchClient = await _baseService.GetIndexClientForAdmin(indexName);
-            IndexDocumentsResult result = searchClient.IndexDocuments(batch);
-            return !(result.Results.Any(r => r.Succeeded == false) == true);
+                var searchClient = await _baseService.GetIndexClientForAdmin(indexName);
+                IndexDocumentsResult result = searchClient.IndexDocuments(batch);
+                return !(result.Results.Any(r => r.Succeeded == false) == true);
+            }
+            else
+            {
+                return false;
+            }
         }
 
         public async Task<bool> DeleteIndex(string indexName)
